+++ src/plugins_types.c
@@ -736,106 +736,6 @@ cleanup:
 }
 
 /**
- * @brief Validate, canonize and store value of the YANG built-in binary type.
- *
- * Implementation of the ly_type_store_clb.
- */
-static LY_ERR
-ly_type_store_binary(const struct ly_ctx *ctx, const struct lysc_type *type, const char *value, size_t value_len,
-        uint32_t options, LY_PREFIX_FORMAT UNUSED(format), void *UNUSED(prefix_data), uint32_t hints,
-        const struct lysc_node *UNUSED(ctx_node), struct lyd_value *storage, struct lys_glob_unres *UNUSED(unres),
-        struct ly_err_item **err)
-{
-    LY_ERR ret = LY_SUCCESS;
-    size_t base64_start, base64_end, base64_count, base64_terminated, value_end;
-    struct lysc_type_bin *type_bin = (struct lysc_type_bin *)type;
-
-    LY_CHECK_ARG_RET(ctx, value, LY_EINVAL);
-
-    /* check hints */
-    ret = ly_type_check_hints(hints, value, value_len, type->basetype, NULL, err);
-    LY_CHECK_GOTO(ret != LY_SUCCESS, cleanup);
-
-    /* validate characters and remember the number of octets for length validation */
-    /* silently skip leading whitespaces */
-    base64_start = 0;
-    while (base64_start < value_len && isspace(value[base64_start])) {
-        base64_start++;
-    }
-    /* silently skip trailing whitespace */
-    value_end = value_len;
-    while (base64_start < value_end && isspace(value[value_end - 1])) {
-        value_end--;
-    }
-
-    /* find end of base64 value */
-    base64_end = base64_start;
-    base64_count = 0;
-    while ((base64_end < value_len) &&
-            /* check correct character in base64 */
-            ((('A' <= value[base64_end]) && (value[base64_end] <= 'Z')) ||
-            (('a' <= value[base64_end]) && (value[base64_end] <= 'z')) ||
-            (('0' <= value[base64_end]) && (value[base64_end] <= '9')) ||
-            ('+' == value[base64_end]) ||
-            ('/' == value[base64_end]) ||
-            ('\n' == value[base64_end]))) {
-
-        if ('\n' != value[base64_end]) {
-            base64_count++;
-        }
-        base64_end++;
-    }
-
-    /* find end of padding */
-    base64_terminated = 0;
-    while (((base64_end < value_len) && (base64_terminated < 2)) &&
-            /* check padding on end of string */
-            (('=' == value[base64_end]) ||
-            ('\n' == value[base64_end]))) {
-
-        if ('\n' != value[base64_end]) {
-            base64_terminated++;
-        }
-        base64_end++;
-    }
-
-    /* check if value is valid base64 value */
-    if (value_end != base64_end) {
-        ret = ly_err_new(err, LY_EVALID, LYVE_DATA, NULL, NULL, "Invalid Base64 character (%c).", value[base64_end]);
-        goto cleanup;
-    }
-
-    if ((base64_count + base64_terminated) & 3) {
-        /* base64 length must be multiple of 4 chars */
-        ret = ly_err_new(err, LY_EVALID, LYVE_DATA, NULL, NULL, "Base64 encoded value length must be divisible by 4.");
-        goto cleanup;
-    }
-
-    /* check if value meets the type requirments */
-    if (type_bin->length) {
-        const uint32_t value_length = ((base64_count + base64_terminated) / 4) * 3 - base64_terminated;
-        ret = ly_type_validate_range(LY_TYPE_BINARY, type_bin->length, value_length, value, err);
-        LY_CHECK_GOTO(ret != LY_SUCCESS, cleanup);
-    }
-
-    if (base64_count != 0) {
-        ret = lydict_insert(ctx, &value[base64_start], base64_end - base64_start, &storage->canonical);
-        LY_CHECK_GOTO(ret != LY_SUCCESS, cleanup);
-    } else {
-        ret = lydict_insert(ctx, "", 0, &storage->canonical);
-        LY_CHECK_GOTO(ret != LY_SUCCESS, cleanup);
-    }
-    storage->ptr = NULL;
-    storage->realtype = type;
-
-cleanup:
-    if (options & LY_TYPE_STORE_DYNAMIC) {
-        free((char *)value);
-    }
-    return ret;
-}
-
-/**
  * @brief Validate and store value of the YANG built-in string type.
  *
  * Implementation of the ly_type_store_clb.
@@ -2143,6 +2043,11 @@ ly_type_free_union(const struct ly_ctx *ctx, struct lyd_value *value)
         value->subvalue = NULL;
     }
 }
+
+/* plugins_types_binary.c */
+extern LY_ERR ly_type_store_binary(const struct ly_ctx *ctx, const struct lysc_type *type,
+        const char *value, size_t value_len, uint32_t options, LY_PREFIX_FORMAT format, void *prefix_data, uint32_t hints,
+        const struct lysc_node *ctx_node, struct lyd_value *storage, struct lys_glob_unres *unres, struct ly_err_item **err);
 
 /* plugins_types_integer.c */
 extern LY_ERR ly_type_store_int(const struct ly_ctx *ctx, const struct lysc_type *type,
